refactor(MetadataMultiInput): rename component to reflect its specific purpose
