+++ app/actions/mappingFuncs.ts
@@ -19,6 +19,14 @@ export function mapVersionInfo (data: any[]): VersionInfo[] {
 }
 
 export function mapHistory (data: any[]): VersionInfo[] {
+  /**
+   * this odd mapping function is the result of changes to the backend api.
+   * history returns a list of `DatasetLogItems`, which contains a `VersionInfo`
+   * with the added fields `commitMessage` and `commitTitle`.
+   * Desktop uses versionInfo in multiple places, and it is easier to mold the
+   * api output to a flatter form we can more easily use. This function flattens
+   * the structure and removes the embedded versionInfo field.
+   */
   return data.map((item: any) => {
     const vinfo = {
       ...item,
